backend/requirements.txt
Django==6.0.1
djangorestframework==3.17.1
django-cors-headers==4.9.0
celery==5.6.3
redis==7.4.0
cryptography==46.0.7
Pillow==12.2.0
python-dotenv==1.2.2
boto3==1.40.38
pywin32==311; platform_system == "Windows"
pypdf==6.10.2
python-docx==1.2.0
